Add unit testing capability
The source file (jinks.c) was added to house any library-wide code.
Additionally, if the macro JX_TESTING is defined, the source file provides a
few functions for automatically running unit tests in all the source files. A
unit test is considered to be any function with the following signature:

  bool jx_test_foo_bar();

To support this, the Makefile has been modified so that (using grep and sed)
it extracts the list of all such functions from all the source files and then
prepares a file called 'test_list.in' which is included into jinks.c when
JX_TESTING is defined.

The recommended approach is then to write unit tests as non-static functions
in the module with the specified signature and to place all the unit tests at
the end of the source file inside a #ifdef JX_TESTING block. This way, unit
tests are compiled out of the final library, but can be compiled into a test
executable (jinks_test) and run. A unit test should return true if it passed.

Four bogus unit tests (all of which fail) were added into jx_buffer.c to
demonstrate how unit tests should be defined and to verify that the framework
compiles and runs all unit tests.