Improve error reporting in unit testing
This update changes the return type of unit tests to jx_test, which is a type
that includes file name, line number, the failed assertion expression, and a
custom error message. Using this special return type, the name of a unit test
is irrelevant and any function with no argument list and a return type of
jx_test is now treated as a unit test by the Makefile script that generates
the unit test harness.

Additionally, the unit test harness was improved and refactored as was the
script that generates the unit tests. Finally, the implementation of jx_buffer
was finished and several legitimate unit tests were written. These tests serve
to verify both the code and the unit test framework.