Add a slice module
A slice object is an indexing interface on top of a shared array of equally
sized items. When a slice is created, the user specifies the number and size
of items in the underlying array. From an existing slice, a new slice may be
created by specifying a starting index, step size, and a count. The new slice
appears to be another array whose first element is array[start], next element
is array[start + step], for up to count elements. If count is too large, the
reslice method reduces it to the maximum number of items that lie in the range
of the slice it is being built on.

For example, if a user created a slice of 20 integers and initialized them to
be [0, 19], then resliced with start=0, step=3, count=20, they would a slice
that represents the array/list [0, 3, 6, 9, 12, 15, 18], that is, all the
items in the original list divisible by 3. Taking a slice on top of that with
start = 0, step = 5, count = 20 yields [0, 15], that is the items divisible by
3*5=15.

Using negative step sizes can reverse the order of the list. The 'start'
argument is permitted to be any integer in the interval [-length, length-1].
Negative start values indicate distance from the end. For example, starting
from the original slice [0..19], a slice of start = -1, step = -1, count = 20,
gives the slice [19,18,...0], that is, it reverses the list.

Of course, the slices just keep a pointer to the underlying array together
with start, stride, and count parameters to correctly index into the
underlying array. Creating a new slice is achieved by computing new start,
stride, and count values.

Since the user never has direct access to the underlying array, and since the
array must remain until all the slices are destroyed, the slices maintain
access to the array using the jx_pointer object that maintains reference
counts.