Modify lib build to include release compile flags
There is a new target 'lib' in the Makefile which just builds the jinks.a
static library. This is constructed from release builds of the sources that
define NDEBUG (to turn off all the asserts), does not define JX_TESTING (so
there is no overhead from unit tests), and uses optimization level O3.

This resolves issue #2.