Add error handling and error return codes
The new enum type jx_result specifies the possible error return codes for the
library. All functions that may have runtime errors (currently, functions that
attempt to allocate memory) now return a value of this type that indicates
whether they succeeded.

A new function in jinks.h provides a single point of access for allocating and
resizing memory blocks which is careful to preserve the pointer, uses NULL
blocks whenever the requested size is zero, and return the out-of-memory code
if the allocation fails.

Furthermore, there are two new macros defined in jinks.h: JX_TRY and
JX_FREE_AND_NULL. The latter is obvious: free a block of memory and
simultaneously NULL the pointer. The former takes an expression (usually a
function call) that returns a jx_result, calls it, checks to see if it
succeeded, and returns the error code if it failed. This makes error handling
easier since functions can do this like:

  JX_TRY(jx_alloc(&array, sz));

And then have all the following code safely assume that the memory allocation
succeeded. If it does fail, the function will return early, which is a
potential downside. But the TRY nomenclature seems common enough.

This resolves issue #3, which is the last of the 0.1 milestone.