Remove buffer module
In retrospect, it seems a bit silly to have a buffer module. The buffer object
did very little, except for resizing a block of memory. But it meant that you
had to store that extra object and make all calls through it. In any case, a
buffer was essentially just a vector where the items are chars. So, I killed
the buffer module and put the functionality directly in the vector.

I also removed some macros I was using for simple stuff and the jx_alloc
function. I don't want this project to get too "frameworky".