Change unit testing to use googletest
Removed custom unit testing framework and replaced it with googletest. This
means that unit tests are written in C++ rather than C, but also has the
advantage of a full-featured unit test framework that people are accustomed
to.

Additionally, the Makefile has been modified so that running 'make' or 'make
all' (the default target is 'all') builds libjinks.a and libjinks-debug.a and
copies them into the root directory. Therefore, the build now produces useable
libraries for both a debug and a release build and puts them in the root.
Consuming the libraries is as simple as including jinks.h and linking to the
correct static library, provided that the include path and library path
include the build location.

At some future point, when I get close to a full release, I may want to add an
install target that puts these files into an appropriate system include and
lib location.