Implement and Test vector implementation of list
After refactoring and restructing the layout of the library, I have now
provided the first implementation of a jx_list using a vector. The
implementation includes some general purpose memory allocation functions and I
wrote fairly extensive unit tests for the module.

The unit tests are setup using googletest's parametrized test capability so
that by adding other init functions to the list in tests/jx_list.cc, all of
the unit tests that were written for the vector implementation will be applied
to all other implementations. This way, a test can be written to define how
the interface should behave and by adding initializers, the test will be
applied to every implementation of the interface to ensure they work the same
way.